Simple Random Sampling

simple_rs implements a random sampling procedure in which units are independently sampled. Because units are sampled independently, the number of units that are sampled can vary from sample to sample. For most applications in which the number of units in the sampling frame is known in advance, complete_rs is better because the number of units sampled is fixed across sampled.

simple_rs(N, prob = NULL, check_inputs = TRUE)

Arguments

N
The number of units. N must be a positive integer. (required)

prob
prob is the probability of being sampled must be a real number between 0 and 1 inclusive. (optional)

check_inputs
logical. Defaults to TRUE.

Value

A numeric vector of length N that indicates if a unit is sampled (1) or not (0).

Examples

S <- simple_rs(N = 100)
table(S)

#> S
#>  0  1 
#> 46 54 

S <- simple_rs(N = 100, prob = 0.3)
table(S)

#> S
#>  0  1 
#> 69 31