Stratified and Clustered Random Sampling

A random sampling procedure in which units are sampled as clusters and clusters are nested within strata.

strata_and_cluster_rs(strata_var, clust_var, prob = NULL, n = NULL,
  strata_n = NULL, strata_prob = NULL, check_inputs = TRUE)

Arguments

strata_var
A vector of length N that indicates which stratum each unit belongs to.

clust_var
A vector of length N that indicates which cluster each unit belongs to.

prob
Use for a design in which either floor(N_clusters_stratum*prob) or ceiling(N_clusters_stratum*prob) clusters are sampled within each stratum. The probability of being sampled is exactly prob because with probability 1-prob, floor(N_clusters_stratum*prob) clusters will be sampled and with probability prob, ceiling(N_clusters_stratum*prob) clusters will be sampled. prob must be a real number between 0 and 1 inclusive. (optional)

n
Use for a design in which the scalar n describes the fixed number of units to sample in each stratum. This number does not vary across strata.

strata_n
Use for a design in which strata_n describes the number of units to sample within each stratum.

strata_prob
Use for a design in which strata_prob describes the probability of being sampled within each stratum. Differs from prob in that the probability of being sampled can vary across strata.

check_inputs
logical. Defaults to TRUE.

Value

A numeric vector of length N that indicates if a unit is sampled (1) or not (0).

Examples

clust_var <- rep(letters, times = 1:26)

strata_var <- rep(NA, length(clust_var))
strata_var[clust_var %in% letters[1:5]] <- "stratum_1"
strata_var[clust_var %in% letters[6:10]] <- "stratum_2"
strata_var[clust_var %in% letters[11:15]] <- "stratum_3"
strata_var[clust_var %in% letters[16:20]] <- "stratum_4"
strata_var[clust_var %in% letters[21:26]] <- "stratum_5"

table(strata_var, clust_var)

#>            clust_var
#> strata_var   a  b  c  d  e  f  g  h  i  j  k  l  m  n  o  p  q  r  s  t  u  v
#>   stratum_1  1  2  3  4  5  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#>   stratum_2  0  0  0  0  0  6  7  8  9 10  0  0  0  0  0  0  0  0  0  0  0  0
#>   stratum_3  0  0  0  0  0  0  0  0  0  0 11 12 13 14 15  0  0  0  0  0  0  0
#>   stratum_4  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 16 17 18 19 20  0  0
#>   stratum_5  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 21 22
#>            clust_var
#> strata_var   w  x  y  z
#>   stratum_1  0  0  0  0
#>   stratum_2  0  0  0  0
#>   stratum_3  0  0  0  0
#>   stratum_4  0  0  0  0
#>   stratum_5 23 24 25 26

S <- strata_and_cluster_rs(strata_var = strata_var,
                          clust_var = clust_var)

table(S, strata_var)

#>    strata_var
#> S   stratum_1 stratum_2 stratum_3 stratum_4 stratum_5
#>   0         6        27        39        36        70
#>   1         9        13        26        54        71

table(S, clust_var)

#>    clust_var
#> S    a  b  c  d  e  f  g  h  i  j  k  l  m  n  o  p  q  r  s  t  u  v  w  x  y
#>   0  1  2  3  0  0  0  0  8  9 10  0 12 13 14  0  0 17  0 19  0  0 22 23  0 25
#>   1  0  0  0  4  5  6  7  0  0  0 11  0  0  0 15 16  0 18  0 20 21  0  0 24  0
#>    clust_var
#> S    z
#>   0  0
#>   1 26

S <- strata_and_cluster_rs(clust_var = clust_var,
                           strata_var = strata_var,
                           prob = .5)

table(S, clust_var)

#>    clust_var
#> S    a  b  c  d  e  f  g  h  i  j  k  l  m  n  o  p  q  r  s  t  u  v  w  x  y
#>   0  1  0  3  0  5  0  7  8  9  0 11  0 13  0  0  0 17 18  0 20  0 22 23  0 25
#>   1  0  2  0  4  0  6  0  0  0 10  0 12  0 14 15 16  0  0 19  0 21  0  0 24  0
#>    clust_var
#> S    z
#>   0  0
#>   1 26

table(S, strata_var)

#>    strata_var
#> S   stratum_1 stratum_2 stratum_3 stratum_4 stratum_5
#>   0         9        24        24        55        70
#>   1         6        16        41        35        71

S <- strata_and_cluster_rs(clust_var = clust_var,
                           strata_var = strata_var,
                           strata_n = c(2, 3, 2, 3, 2))

table(S, clust_var)

#>    clust_var
#> S    a  b  c  d  e  f  g  h  i  j  k  l  m  n  o  p  q  r  s  t  u  v  w  x  y
#>   0  1  2  3  0  0  0  0  8  0 10  0 12 13  0 15  0 17  0 19  0  0 22 23 24 25
#>   1  0  0  0  4  5  6  7  0  9  0 11  0  0 14  0 16  0 18  0 20 21  0  0  0  0
#>    clust_var
#> S    z
#>   0  0
#>   1 26

table(S, strata_var)

#>    strata_var
#> S   stratum_1 stratum_2 stratum_3 stratum_4 stratum_5
#>   0         6        18        40        36        94
#>   1         9        22        25        54        47

S <- strata_and_cluster_rs(clust_var = clust_var,
                           strata_var = strata_var,
                           strata_prob = c(.1, .2, .3, .4, .5))

table(S, clust_var)

#>    clust_var
#> S    a  b  c  d  e  f  g  h  i  j  k  l  m  n  o  p  q  r  s  t  u  v  w  x  y
#>   0  1  2  3  4  5  6  0  8  9 10 11 12 13 14  0 16  0 18 19  0 21  0  0 24 25
#>   1  0  0  0  0  0  0  7  0  0  0  0  0  0  0 15  0 17  0  0 20  0 22 23  0  0
#>    clust_var
#> S    z
#>   0  0
#>   1 26

table(S, strata_var)

#>    strata_var
#> S   stratum_1 stratum_2 stratum_3 stratum_4 stratum_5
#>   0        15        33        50        53        70
#>   1         0         7        15        37        71